[Sqlparser] Pool the yyParserImpls
For a high QPS application, the sqlparser generates a ton of garbage.
```
622GB    26.33% 60.87%   880.50GB 37.28% vitess.io/vitess/go/vt/sqlparser.yyParse
212.57GB  9.00% 69.87%   258.49GB 10.94% vitess.io/vitess/go/vt/sqlparser.(*yyParserImpl).Parse
```

Most of this, per the heap profile is allocations of new `yyParserImpl`
objects for every call to `yyParse`. This diff uses `sync.Pool` to
mitigate the total allocation overhead.

Modifies sql.y to fix a place where we took a reference to a stack
variable rather than making an intermediate copy.

Resolves https://github.com/vitessio/vitess/issues/4174

Signed-off-by: Daniel Tahara <tahara@dropbox.com>
Signed-off-by: Shlomi Noach <shlomi-noach@github.com>